refactor(web/stores): split websocket into transport/dispatch/subscriptions package
Decomposes the 818-line stores/websocket.ts monolith into a sibling-
aggregator package mirroring stores/connections.ts and stores/providers.ts.

- stores/websocket.ts (44 lines): aggregator. Composes the transport
  and subscriptions slices into a Zustand store; teardown() composes
  the two slice-level teardown helpers and resets observable state.
  Vite HMR dispose hook preserved.
- stores/websocket/types.ts: WebSocketState interface plus WsSet/WsGet
  type aliases.
- stores/websocket/dispatch.ts: pure parse helpers (isWsEvent,
  eventVersion, isWsChannelArray, estimateByteLength); no module
  state.
- stores/websocket/subscriptions.ts: owns channelHandlers, active and
  pending subscription bookkeeping. Exports dispatchEvent,
  replaySubscriptions, and teardownSubscriptions.
- stores/websocket/transport.ts: owns the live socket and heartbeat,
  pong, and reconnect timer handles. Action methods implemented as
  top-level functions taking (set, get), so the slice factory body
  itself stays well under the 80 LPF cap. Per-stage helpers extracted
  to keep every function complexity at most 8.
- stores/websocket/sse-fallback.ts: SSE-fallback transport
  bookkeeping. Owns sseClient and proxy-block suspicion counter.
- stores/websocket/transport-shared.ts: tiny getCurrentSocket and
  setCurrentSocket pair that breaks the otherwise circular
  transport-subscriptions module load.

Every new file under 400 LOC, every function under 80 LPF, every
function complexity at most 8. All 36 existing websocket tests pass
unchanged (the file-local resetStore() and teardown() carve-out from
test-setup.tsx is preserved). The full 600-test stores suite stays
green with no active-handle leaks. WS_PROTOCOL_VERSION continues to
flow from utils/constants.ts (no literal numbers).

Part of #2093 (PR B of EPIC #2066).
